[Feat/#304] :wrench: Refactor: 결과 페이지 모든 문제 한 번에 보여주는 기획으로 수정
