--- chatty-be/src/main/java/com/chatty/chatty/player/domain/entity/MarkedAnswer.java
@@ -1,10 +1,0 @@
-package com.chatty.chatty.player.domain.entity;
-
-import lombok.Builder;
-
-@Builder
-public record MarkedAnswer(
-        String playerAnswer,
-        Boolean marking
-) {
-}
