+++ chatty-be/src/main/java/com/chatty/chatty/player/domain/entity/Player.java
@@ -11,8 +11,6 @@ import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
-import java.util.ArrayList;
-import java.util.List;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -44,8 +42,4 @@ public class Player extends BaseEntity {
 
     @Column(length = 10, nullable = false)
     private String nickname;
-
-    @Column(nullable = false)
-    @Builder.Default
-    private List<MarkedAnswer> answers = new ArrayList<>();
 }
