+++ chatty-be/src/main/java/com/chatty/chatty/quizroom/controller/dto/RoomResultResponse.java
@@ -1,10 +1,25 @@
 package com.chatty.chatty.quizroom.controller.dto;
 
+import java.util.List;
+
 public record RoomResultResponse(
-        Integer quizNumber,
-        String quiz,
-        String playerAnswer,
-        Boolean marking,
-        Integer correctRate
+        List<QuizResult> results
 ) {
+    public record QuizResult(
+            Integer quizNumber,
+            String quiz,
+            List<String> options,
+            String correct,
+            List<PlayerAnswer> playerAnswers,
+            Integer correctRate
+    ) {
+        public record PlayerAnswer(
+                Long playerId,
+                String nickname,
+                String playerAnswer,
+                Boolean marking
+        ) {
+        }
+    }
 }
+
